requirements.txt
Django>=5.2
djangorestframework>=3.16
psycopg2>=2.9
django-filter>=23.0
django-cors-headers>=4.0.0
python-decouple>=3.8
gunicorn>=21.0.0
whitenoise>=6.0.0
django-environ>=0.12.0
boto3>=1.40.33
scikit-learn>=1.7
numpy>=2.3
pandas>=2.3
torch>=2.8.0
sentence-transformers>=5.1.1
mlflow>=2.8
python-dotenv>=1.1.1